+++ homes/notashelf/programs/terminal/editors/neovim/lua/core.lua
@@ -1,12 +1,6 @@
 local options = {
 	-- disable the -- STATUS -- line
 	showmode = false,
-
-	-- spellchecking
-	spell = true,
-
-	-- spell langs
-	spelllang = { "en" },
 }
 
 
@@ -15,11 +9,3 @@ local options = {
 for key, value in pairs(options) do
 	vim.opt[key] = value
 end
-
-vim.api.nvim_create_autocmd({ "FileType" }, {
-	pattern = { "toggleterm" },
-	callback = function()
-		vim.opt_local.wrap = false
-		vim.opt_local.spell = false
-	end,
-})