+++ homes/notashelf/programs/terminal/editors/neovim/lua/neovide.lua
@@ -1,5 +1,0 @@
-if vim.g.neovide then
-	vim.o.guifont = "Iosevka Nerd Font:h14" -- text below applies for VimScript
-	vim.g.neovide_theme = 'auto'
-	vim.g.neovide_transparency = 0.95
-end